fix(dev): CTL-1914/1918 — lint clean, and extract install-completeness so doctor.mjs stays reviewable
- shellcheck: SC2034 (CATALYST_SOURCE_ORIGIN/_REASON are read by callers and by the
  test — exported, which is also how shellcheck can see the use), a dead SCRIPT_DIR in
  setup_sweep_config left over from the resolver rewrite, SC2015 (`A && pass || fail`
  runs the failure branch when `pass` itself exits non-zero, so an assertion could
  report a failure it never had), SC2115 (`${SCRATCH:?}` guards), SC2016 (the
  single-quoted run_fn bodies are source for a CHILD shell — declared once at file head,
  since a mid-file directive only covers the next command).

- checkInstallCompleteness moves to its own module + a zero-import doctor-status.mjs
  leaf holding STATUS/mkCheck. doctor.mjs is NOT prettier-formatted on main — verified
  against a pristine origin/main checkout, not assumed — so any edit to it drags a
  ~1,100-line reformat into the diff and buries a 100-line change in 1,000 lines of
  someone else's code. Extraction keeps doctor.mjs's delta at 6 lines. STATUS and
  mkCheck are re-exported from doctor.mjs, so every existing importer is unchanged.

trunk check: the only remaining unformatted file is doctor.mjs, which is unformatted on
main too (control run above). Left alone deliberately.
