+++ plugins/dev/scripts/execution-core/doctor.mjs
@@ -27,6 +27,11 @@
 import { readFileSync, statSync, existsSync, lstatSync, realpathSync, readdirSync, openSync, readSync, closeSync } from "node:fs";
 import { resolve, dirname, isAbsolute, join } from "node:path";
 import { homedir } from "node:os";
+// CTL-1918: install-completeness lives in its own module. doctor.mjs is NOT
+// prettier-formatted on main (verified against a pristine checkout), so any edit here
+// drags a ~1,100-line reformat into the diff — extracting keeps this change reviewable.
+import { STATUS, mkCheck } from "./doctor-status.mjs";
+import { checkInstallCompleteness } from "./install-completeness.mjs";
 import { fileURLToPath } from "node:url";
 import { spawnSync, execFileSync } from "node:child_process";
 
@@ -153,9 +158,13 @@ function readLinearBotUserIds(l1Path, l2Path) {
 
 // ─── Check model ─────────────────────────────────────────────────────────────
 
-export const STATUS = { PASS: "pass", WARN: "warn", FAIL: "fail", INFO: "info" };
+// CTL-1918: STATUS/mkCheck moved to the zero-import leaf doctor-status.mjs so a check
+// can live in its own module without importing this one (which would be circular).
+// Re-exported here, so every existing `import { STATUS, mkCheck } from "./doctor.mjs"`
+// keeps working unchanged.
+export { STATUS, mkCheck } from "./doctor-status.mjs";
 
-export const mkCheck = (name, status, detail) => ({ name, status, detail });
+export { checkInstallCompleteness };
 
 // ─── CTL-1616 PR2/PR3: secret-contract observability (zero grade change) ─────
 //
@@ -3931,108 +3940,6 @@ export function checkDrainDisabled(deps = {}) {
     "drain-disabled",
     STATUS.INFO,
     "not drain-disabled — this node honors the drain flag (CTL-1678)",
-  );
-}
-
-// checkInstallCompleteness — CTL-1918. Does this node carry the four things a
-// FINISHED install leaves behind, or only the things a STARTED one does?
-//
-// setup-catalyst.sh used to end by PRINTING instructions for steps it could perform,
-// so "setup completed successfully" and "this node works" were different states with
-// nothing measuring the gap. setup now performs them and records what it could not —
-// but that ledger is printed once, at the end of a run nobody re-reads. This is the
-// standing answer: the same four outcomes, checkable at any time.
-//
-// ⛔ ADVISORY — never FAIL. Install shape is an operator repair, and doctor's FAIL
-// count gates worker activation; a half-installed node must not be made unable to
-// work on the strength of a cosmetic gap. Same posture, and same reason, as
-// checkRegistryTeamIdentity.
-//
-// ⚠️ Every leg is three-valued. "I could not look" (unreadable config, absent
-// registry, a platform with no launchd) is reported as UNKNOWN and never folded into
-// "absent" — a check that reports a missing step for a file it failed to open is how
-// an operator gets sent to repair something that was never broken.
-export function checkInstallCompleteness(deps = {}) {
-  const {
-    env = process.env,
-    home = homedir(),
-    exists = (p) => existsSync(p),
-    readJson = (p) => {
-      try {
-        return { ok: true, value: JSON.parse(readFileSync(p, "utf8")) };
-      } catch {
-        return { ok: false, value: null };
-      }
-    },
-    layer2 = layer2Path(),
-    registryPath = join(home, "catalyst", "execution-core", "registry.json"),
-    platform = process.platform,
-  } = deps;
-
-  const legs = [];
-  const note = (name, state, detail) => legs.push({ name, state, detail });
-
-  // 1. the catalyst-* CLIs on PATH — what install-cli.sh leaves behind.
-  const binDir = env.CATALYST_BIN_DIR || env.CATALYST_CLI_BIN_DIR || join(home, ".catalyst", "bin");
-  note("cli", exists(join(binDir, "catalyst-stack")) ? "ok" : "missing", binDir);
-
-  // 2. pluginDirs registered — what setup-plugin-source.sh leaves behind. An
-  //    unreadable Layer-2 file is UNKNOWN: the key may well be there.
-  const cfg = readJson(layer2);
-  if (!cfg.ok) {
-    note("plugin-source", "unknown", `Layer-2 config unreadable (${layer2})`);
-  } else {
-    const pd = cfg.value?.catalyst?.orchestration?.pluginDirs;
-    const first = Array.isArray(pd) ? pd[0] : typeof pd === "string" ? pd : null;
-    note("plugin-source", first ? "ok" : "missing", first || "no pluginDirs key");
-  }
-
-  // 3. the orphan-sweep LaunchAgent — macOS only. On any other platform this is not
-  //    "missing", it is not applicable, and saying "missing" would be a false repair.
-  if (platform !== "darwin") {
-    note("orphan-sweep", "unknown", "not macOS — launchd scheduling is a follow-up (CTL-1030)");
-  } else {
-    const plist = join(home, "Library", "LaunchAgents", "ai.coalesce.catalyst-orphan-sweep.plist");
-    note("orphan-sweep", exists(plist) ? "ok" : "missing", plist);
-  }
-
-  // 4. at least one project enrolled — without a registry entry the daemon dispatches
-  //    nothing, which looks exactly like a broken one.
-  const reg = readJson(registryPath);
-  if (!exists(registryPath)) {
-    note("registry", "missing", `no registry at ${registryPath}`);
-  } else if (!reg.ok) {
-    note("registry", "unknown", `registry unreadable (${registryPath})`);
-  } else {
-    const projects = reg.value?.projects;
-    const n = Array.isArray(projects) ? projects.length : projects && typeof projects === "object" ? Object.keys(projects).length : 0;
-    note("registry", n > 0 ? "ok" : "missing", `${n} project(s) enrolled`);
-  }
-
-  const missing = legs.filter((l) => l.state === "missing");
-  const unknown = legs.filter((l) => l.state === "unknown");
-  const fmt = (ls) => ls.map((l) => `${l.name} (${l.detail})`).join("; ");
-
-  if (missing.length === 0 && unknown.length === 0) {
-    return mkCheck(
-      "install-completeness",
-      STATUS.PASS,
-      "install is complete — CLIs on PATH, plugin-source registered, orphan-sweep scheduled, project enrolled",
-    );
-  }
-  if (missing.length === 0) {
-    return mkCheck(
-      "install-completeness",
-      STATUS.INFO,
-      `install looks complete, but ${unknown.length} leg(s) could not be measured: ${fmt(unknown)}`,
-    );
-  }
-  return mkCheck(
-    "install-completeness",
-    STATUS.WARN,
-    `install is INCOMPLETE — ${missing.length} step(s) never landed: ${fmt(missing)}` +
-      (unknown.length ? ` | unmeasured: ${fmt(unknown)}` : "") +
-      " — re-run setup-catalyst.sh, or complete them from its deferred-step list (CTL-1918)",
   );
 }
